+++ mayan/apps/documents/handlers.py
@@ -9,7 +9,6 @@ from .settings import (
     setting_document_version_page_image_cache_maximum_size
 )
 from .signals import signal_post_initial_document_type
-from .tasks import task_duplicates_clean_empty_lists, task_duplicates_scan_for
 
 
 def handler_create_default_document_type(sender, **kwargs):
@@ -42,13 +41,3 @@ def handler_create_document_version_page_image_cache(sender, **kwargs):
             'maximum_size': setting_document_version_page_image_cache_maximum_size.value,
         }, defined_storage_name=STORAGE_NAME_DOCUMENT_VERSION_PAGE_IMAGE_CACHE,
     )
-
-
-def handler_scan_duplicates_for(sender, instance, **kwargs):
-    task_duplicates_scan_for.apply_async(
-        kwargs={'document_id': instance.document_id}
-    )
-
-
-def handler_remove_empty_duplicates_lists(sender, **kwargs):
-    task_duplicates_clean_empty_lists.apply_async()