+++ mayan/apps/documents/icons.py
@@ -284,19 +284,13 @@ icon_document_version_page_zoom_out = Icon(
     driver_name='fontawesome', symbol='search-minus'
 )
 
-# Duplicated document
-
-icon_duplicated_document_list = Icon(
-    driver_name='fontawesome', symbol='clone'
-)
-icon_duplicated_document_scan = Icon(
-    driver_name='fontawesome-dual-classes', primary_class='fas fa-clone',
-    secondary_class='far fa-eye'
-)
+# Recently created
 
 icon_document_recently_created_list = Icon(
     driver_name='fontawesome', symbol='asterisk'
 )
+
+# Trashed documents
 
 icon_trash_can_empty = Icon(
     driver_name='fontawesome-dual-classes', primary_class='fas fa-trash-alt',