+++ mayan/apps/documents/models/__init__.py
@@ -4,7 +4,6 @@ from .document_file_page_models import *  # NOQA
 from .document_version_models import *  # NOQA
 from .document_version_page_models import *  # NOQA
 from .document_type_models import *  # NOQA
-from .duplicated_document_models import *  # NOQA
 from .favorite_document_models import *  # NOQA
 from .recently_accessed_document_models import *  # NOQA
 from .trashed_document_models import *  # NOQA