+++ mayan/apps/duplicates/tests/mixins.py
@@ -6,10 +6,10 @@ class DuplicatedDocumentsTestMixin:
 class DuplicatedDocumentsViewsTestMixin:
     def _request_test_document_duplicates_list_view(self):
         return self.get(
-            viewname='documents:document_duplicates_list', kwargs={
+            viewname='duplicates:document_duplicates_list', kwargs={
                 'document_id': self.test_documents[0].pk
             }
         )
 
     def _request_test_duplicated_document_list_view(self):
-        return self.get(viewname='documents:duplicated_document_list')
+        return self.get(viewname='duplicates:duplicated_document_list')